feat(DataTable): add `wdith` adjustment to columns
- Have columns be able to set
column width and if not set width should auto adjust itself

Refs: #3
